+++ src/components/mbx-button/index.js
@@ -1,0 +1,3 @@
+import main from './mbx-button';
+
+export default main;